refactor(stint-auction): do not use unnecessary generator function
Changelog: do not use unnecessary generator function
