+++ packages/widgets/stint-auction/src/components/StintAuction/formModifier.js
@@ -1,0 +1,16 @@
+import {form} from 'tocco-app-extensions'
+
+export function modifyFormDefinition(formDefinition, appContext) {
+  if (formDefinition.id === 'Stint_auction_detail') {
+    // allows action to read widget config we're on
+    return form.adjustActions(formDefinition, 'stintAuctionRegisterLecturer', action => ({
+      ...action,
+      properties: {
+        ...action.properties,
+        widgetKey: appContext.widgetConfigKey
+      }
+    }))
+  }
+
+  return formDefinition
+}
